Implementación para la consulta de pasantes.

+++ Modulo UnidadDeportiva/Controllers/AuxiliarController.cs
@@ -63,7 +63,7 @@ namespace Modulo_UnidadDeportiva.Controllers
 
             if (model.codEstudiante!=null)
             {
-                model = _pasante.GetAsistenciaPasante((int)model.codEstudiante);
+                model = _pasante.GetAsistenciaPasante(model.codEstudiante);
             }
             return View(model);
         }
